+++ R/sm_slope.R
@@ -4,12 +4,9 @@
 #' This function generates a slope chart.
 #' This is very useful for comparing the effect between two time points.
 #'
-#' For this function to work properly,
-#' ggplot()'s mapping has to be quite specific.
-#' 1. Each observation has to be grouped.
+#' ggplot()'s mapping has to be quite specific: each observation has to be grouped.
 #'
-#' With smplot2, average plot can be displayed along with error bars, both of which
-#' are optional.
+#' Error bar types can be specified (ci, sd, and se).
 #'
 #' @param ...
 #' List of parameters for individual points and lines across different elements
@@ -67,7 +64,7 @@
 #' If the legend is not needed, the input should be FALSE.
 #'
 #'
-#' @return Returns a double vector output representing the slope from linear regression.
+#' @return Returns a slope chart which is a ggplot2 object.
 #' @import ggplot2 cowplot Hmisc
 #' @importFrom stats sd
 #' @importFrom utils modifyList
@@ -97,7 +94,7 @@
 #'                              alpha = 0.3),
 #'           avgPoint.params = list(color='transparent', shape = 21,
 #'                                  size = 4, fill = sm_color('blue')),
-#'           avgLine.params = list(color = sm_color('blue'), size = 1),
+#'           avgLine.params = list(color = sm_color('blue'), linewidth = 1),
 #'           show_mean = TRUE)
 #'
 #'
@@ -105,7 +102,7 @@ sm_slope <- function(...,
                      labels,
                      group,
                      line.params = list(color = 'gray53',
-                                        size = 0.4,
+                                        linewidth = 0.4,
                                         alpha = 0.4),
                      point.params = list(size = 2.5, shape = 21,
                                          color = 'white'),