+++ R/sm_statCorr.R
@@ -1,5 +1,5 @@
-#' Linear regression slope and statistical values
-#' from a paired correlation test
+#' Linear regression slope and statistical values (R or R2 and p values)
+#' from a paired correlation test.
 #'
 #' @description
 #' This combines two different functions:
@@ -53,6 +53,10 @@
 #' If the legend needs to be displayed, the input should be TRUE.
 #' If the legend is not needed, the input should be FALSE.
 #'
+#' @param r2
+#' FALSE or TRUE. TRUE if user wants to compute R2.
+#' @param R2
+#' Same as r2.
 #'
 #' @return Plots a best-fitted linear regression on a correlation plot
 #' with results from correlation statistical tests.
@@ -66,7 +70,11 @@
 #' library(ggplot2)
 #' ggplot(data = mtcars, mapping = aes(x = drat, y = mpg)) +
 #' geom_point(shape = 21, fill = '#0f993d', color = 'white', size = 3) +
-#'  sm_statCorr()
+#'  sm_statCorr() # computes R
+#'
+#'  ggplot(data = mtcars, mapping = aes(x = drat, y = mpg)) +
+#' geom_point(shape = 21, fill = '#0f993d', color = 'white', size = 3) +
+#'  sm_statCorr(R2 = TRUE) # computes R2
 #'
 sm_statCorr <- function(...,
                         fit.params = list(),
@@ -78,7 +86,12 @@ sm_statCorr <- function(...,
                         text_size = 4,
                         show_text = TRUE,
                         borders = TRUE,
-                        legends = FALSE) {
+                        legends = FALSE,
+                        r2 = FALSE, R2) {
+
+  if (!missing(R2)) {
+    r2 <- R2
+  }
 
   params <- list(...)
   fit.params <- modifyList(params, fit.params)
@@ -87,13 +100,32 @@ sm_statCorr <- function(...,
                      modifyList(list(method = 'lm', se = FALSE,
                                      alpha = 0.2, weight = 0.8), fit.params))
 
+  if (r2 == FALSE) { # R
+    textPlot <- ggpubr::stat_cor(p.accuracy = 0.001, method = corr_method,
+                                 alternative = alternative,
+                                 label.sep = separate_by,
+                                 label.x = label_x,
+                                 label.y = label_y,
+                                 size = text_size)
+  } else { #R2
+    if (separate_by == "\n") {
+      textPlot <- ggpubr::stat_cor(aes(label = paste(paste0('atop(',after_stat(rr.label),',', after_stat(p.label),')'),
+                                                     sep = "~~")),
+                                   p.accuracy = 0.001, method = corr_method,
+                                   alternative = alternative,
+                                   label.x = label_x,
+                                   label.y = label_y,
+                                   size = text_size)
+    } else { textPlot <- ggpubr::stat_cor(aes(label = paste(after_stat(rr.label), after_stat(p.label),
+                                                            sep = paste0("~`",separate_by,"`~"))),
+                                          p.accuracy = 0.001, method = corr_method,
+                                          alternative = alternative,
+                                          label.x = label_x,
+                                          label.y = label_y,
+                                          size = text_size)
+    }
+  }
 
-  textPlot <- ggpubr::stat_cor(p.accuracy = 0.001, method = corr_method,
-                               alternative = alternative,
-                               label.sep = separate_by,
-                               label.x = label_x,
-                               label.y = label_y,
-                               size = text_size)
 
   if (show_text == FALSE) {
     textPlot <- NULL
@@ -103,3 +135,5 @@ sm_statCorr <- function(...,
        textPlot,
        sm_hvgrid(borders = borders, legends = legends))
 }
+
+globalVariables(c('rr.label', 'p.label'))